Cleanup concats and start on traffic
The concats are all the same, so they can be lumped together, enhancing
brevity.  Include a templated shorewall.conf (need to add shorewall6.conf) and
modify iface a tad to support the tcinterfaces file.  This makes it possible to
define bandwidth limits on interfaces that are subject to them.  TODO: tcrules,
tcpri.