Restructured the module
Turn shorewall::multi into what it really is... the shorewall class.  Also,
move a bunch of defines to eliminate the 'multi' component of their names.
Since nothing else uses shorewall::base, and all it does is declare the
packages through some cumbersome abstractions, get rid of most of that too and
replace it with resource declarations in the main class.