Puppet 4, module data, and mangle support
Update the module to support puppet 4, including a transition from `Modulefile`
to `metadata.json`, fixups of file and concat mode specification, and reworking
of fact references.

Get rid of the `shorewall_version` fact and instead of using conditional
statements around this fact to select behaviors in the manifests, use hiera in
the form of module data. I think I've got the values set up correctly for
different LTS versions of Ubuntu and released versions of Debian, but no
guarantees.

Finally, create more explicit support for mangle rules, and rework the mark
define as a special case of it.